+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # 0.X.Y - DD/MM/YYYY
 - Fix: Wifi::disconnect would get ssid instead of connection name
 - Fix: In connection name command usages, guard in ticks if it contains spacings
+- Fix: Wifi::set_metric would use connection name instead of ssid
 
 # 0.3.2 - 06/05/2025
 - Fix: Wifi::up